+++ shopelectro/tests/tests_models.py
@@ -1,0 +1,23 @@
+from django.forms.models import model_to_dict
+from django.test import TestCase
+
+from shopelectro.models import Product
+
+
+class ProductModel(TestCase):
+
+    fixtures = ['dump.json']
+
+    def test_creation_deactivated_product(self):
+        """Creation of a deactivated product does not fail."""
+        unactive_product = Product.objects.first()
+        unactive_product.page.is_active = False
+        unactive_product.page.save()
+
+        try:
+            Product.objects.create(**model_to_dict(
+                unactive_product,
+                ['name', 'price', 'vendor_code', 'uuid'],
+            ))
+        except Exception as error:
+            self.fail(f'Creation of existing product failed: {{ error }}')
